+++ bidrag-transport/src/main/kotlin/no/nav/bidrag/transport/behandling/felles/grunnlag/Felles.kt
@@ -99,7 +99,7 @@ interface GrunnlagPeriodeInnhold : GrunnlagInnhold {
 
 interface GrunnlagPeriodeInnholdKopi : GrunnlagInnhold {
     val periode: ÅrMånedsperiode
-    val fraVedtakId: Long
+    val fraVedtakId: Int
 }
 
 interface InnhentetGrunnlagInnhold<out T> : GrunnlagInnhold {
