+++ bidrag-transport/src/main/kotlin/no/nav/bidrag/transport/behandling/statistikk/ForskuddHendelse.kt
@@ -5,7 +5,7 @@ import java.time.LocalDate
 import java.time.LocalDateTime
 
 data class ForskuddHendelse(
-    val vedtaksid: Long,
+    val vedtaksid: Int,
     val vedtakstidspunkt: LocalDateTime,
     val type: String,
     val saksnr: String,
